Refactor mongo-client (for notifications) into \internal and \interfaces (This changeset carries on a merge effort that is under *heavy* construction).
